extract date formatter to utils
